backend/requirements.txt
crewai
fastapi
pydantic
python-dotenv
uvicorn